requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.3.4
aiohttp==3.10.1
aiosignal==1.3.1
aiosqlite==0.20.0
anyio==4.4.0
ariadne==0.23.0
attrs==24.2.0
blinker==1.8.2
click==8.1.7
databases==0.9.0
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
graphql-core==3.2.3
greenlet==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
multidict==6.0.5
python-dateutil==2.9.0.post0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.38.2
strawberry-graphql==0.237.3
typing_extensions==4.12.2
Werkzeug==3.0.3
yarl==1.9.4